refactor(httpClient): index 파일 이름 변경
